Rename models for consistency
