+++ src/api/town/movement.model.ts
@@ -46,4 +46,4 @@ Movement.init({
   },
 }, { sequelize: world.sequelize });
 
-import { Town } from '../town/Town.model';
+import { Town } from '../town/town.model';
