+++ src/api/world/UnitQueue.model.ts
@@ -42,4 +42,4 @@ UnitQueue.init({
   },
 }, { sequelize: world.sequelize });
 
-import { Town } from '../town/Town.model';
+import { Town } from '../town/town.model';
