--- src/api/world/gamez.txt
@@ -1,11 +1,0 @@
-mansions of madness
-imperial settlers
-eclipse
-five tribes
-
-twilight imperium
-
-  nations similar
-  https://www.youtube.com/watch?v=bGBOYu4pMa0
-
-  2nd edition 2018
