+++ src/config/socket.ts
@@ -2,7 +2,7 @@ import * as socketJwt from 'socketio-jwt';
 import config from './environment';
 import WorldData from '../components/world';
 import initializePlayerSocket from '../api/world/player.socket';
-import initializeAllianceSocket from '../api/world/alliance.socket';
+import initializeAllianceSocket from '../api/alliance/alliance.socket';
 import initializeTownSocket from '../api/town/town.socket';
 import initializeMapSocket from '../api/map/map.socket';
 import { logger } from '../';
